Refactor file structure and improve initialization scripts
Moved example files to root directory and adjusted imports accordingly. Streamlined initialization scripts by modifying directory creation logic and adding App Initialization instructions to README. Updated email_downloader to use the updated directory structure.
